Deleted ludmila, add ludmila_v1
