+++ keyboards/mexsistor/ludmila_v1/info.json
@@ -1,5 +1,5 @@
 {
-    "keyboard_name": "ludmila",
+    "keyboard_name": "ludmila_v1",
     "url": "",
     "maintainer": "Kevin M.",
     "width": 3,
