+++ keyboards/mexsistor/ludmila_v1/matrix.c
@@ -256,14 +256,3 @@ static bool read_encoder_switches(matrix_row_t current_matrix[], uint8_t current
 
     return (last_row_value != current_matrix[current_row]);
 }
-void matrix_init_custom(void) {
-    // TODO: initialize hardware here
-}
-
-bool matrix_scan_custom(matrix_row_t current_matrix[]) {
-    bool matrix_has_changed = false;
-
-    // TODO: add matrix scanning routine here
-
-    return matrix_has_changed;
-}
