+++ keyboards/mexsistor/ludmila_v1/readme.md
@@ -1,8 +1,8 @@
-# Ludmila Pad
+# ludmila_v1
 
-![ludmila](https://mexsistor.com/wp-content/uploads/2020/09/IMG_2249-scaled.jpg)
+![ludmila_v1](https://mexsistor.com/wp-content/uploads/2020/09/IMG_2249-scaled.jpg)
 
-Ludmila is a 4 Key Macropad with one rotary encoder.
+Ludmila v1 is a 4 Key Macropad with one rotary encoder.
 
 * Keyboard Maintainer: [Kevin M.](https://github.com/mexsistor)
 * Hardware Supported: Ludmila PCB (atmega32u4)
@@ -10,10 +10,10 @@ Ludmila is a 4 Key Macropad with one rotary encoder.
 
 Make example for this keyboard (after setting up your build environment):
 
-    make mexsistor/ludmila:default
+    make mexsistor/ludmila_v1:default
 
 Flashing example for this keyboard:
 
-    make mexsistor/ludmila:default:flash
+    make mexsistor/ludmila_v1:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
